SOCKS4 proxy support. #77
* dmHttpClient uses socks proxy if env-var DMSOCKS_PROXY is set
* Added jsocks-server for tests

+++ src/test/test_httpclient.cpp
@@ -42,8 +42,8 @@ public:
 
     virtual void TearDown()
     {
-        ASSERT_NE((void*) 0, m_Client);
-        dmHttpClient::Delete(m_Client);
+        if (m_Client)
+            dmHttpClient::Delete(m_Client);
     }
 };
 
@@ -295,6 +295,71 @@ TEST_F(dmHttpClientTest, Test400)
         ASSERT_EQ(dmHttpClient::RESULT_NOT_200_OK, r);
         ASSERT_EQ(400, m_StatusCode);
     }
+}
+
+TEST_F(dmHttpClientTest, InvalidProxy)
+{
+    setenv("DMSOCKS_PROXY", "invalid_host", 1);
+
+    // NOTE: Reconnect with proxy here
+    dmHttpClient::Delete(m_Client);
+    dmHttpClient::NewParams params;
+    params.m_Userdata = this;
+    params.m_HttpContent = dmHttpClientTest::HttpContent;
+    params.m_HttpHeader = dmHttpClientTest::HttpHeader;
+    m_Client = dmHttpClient::New(&params, g_Hostname, 7000);
+    unsetenv("DMSOCKS_PROXY");
+    ASSERT_EQ((void*) 0, (void*) m_Client);
+}
+
+TEST_F(dmHttpClientTest, InvalidProxyPort)
+{
+    setenv("DMSOCKS_PROXY", "localhost", 1);
+    setenv("DMSOCKS_PROXY_PORT", "1082", 1);
+
+    // NOTE: Reconnect with proxy here
+    dmHttpClient::Delete(m_Client);
+    dmHttpClient::NewParams params;
+    params.m_Userdata = this;
+    params.m_HttpContent = dmHttpClientTest::HttpContent;
+    params.m_HttpHeader = dmHttpClientTest::HttpHeader;
+    m_Client = dmHttpClient::New(&params, g_Hostname, 7000);
+
+    dmHttpClient::Result r;
+    r = dmHttpClient::Get(m_Client, "/add/10/20");
+    unsetenv("DMSOCKS_PROXY");
+    unsetenv("DMSOCKS_PROXY_PORT");
+    ASSERT_EQ(dmHttpClient::RESULT_SOCKET_ERROR, r);
+    ASSERT_EQ(dmSocket::RESULT_CONNREFUSED, dmHttpClient::GetLastSocketResult(m_Client));
+}
+
+TEST_F(dmHttpClientTest, SimpleProxy)
+{
+    char buf[128];
+
+    setenv("DMSOCKS_PROXY", "localhost", 1);
+    setenv("DMSOCKS_PROXY_PORT", "1081", 1);
+
+    // NOTE: Reconnect with proxy here
+    dmHttpClient::Delete(m_Client);
+    dmHttpClient::NewParams params;
+    params.m_Userdata = this;
+    params.m_HttpContent = dmHttpClientTest::HttpContent;
+    params.m_HttpHeader = dmHttpClientTest::HttpHeader;
+    m_Client = dmHttpClient::New(&params, g_Hostname, 7000);
+
+    for (int i = 0; i < 100; ++i)
+    {
+        m_Content = "";
+        sprintf(buf, "/add/%d/1000", i);
+        dmHttpClient::Result r;
+        r = dmHttpClient::Get(m_Client, buf);
+        ASSERT_EQ(dmHttpClient::RESULT_OK, r);
+        ASSERT_EQ(1000 + i, strtol(m_Content.c_str(), 0, 10));
+    }
+
+    unsetenv("DMSOCKS_PROXY");
+    unsetenv("DMSOCKS_PROXY_PORT");
 }
 
 #endif // #ifndef _WIN32
